+++ apps/nextjs-website/react-components/components/index.ts
@@ -1,4 +1,5 @@
 import Hero from './Hero/Hero';
 import Editorial from './Editorial/Editorial';
+import HowTo from './HowTo/HowTo';
 
-export { Hero, Editorial };
+export { Hero, Editorial, HowTo };
