--- apps/nextjs-website/react-components/src/index-props.ts
@@ -1,5 +1,0 @@
-import { HeroProps } from './utils/Components.types';
-import { EditorialProps } from './utils/Components.types';
-import { HowToProps } from './utils/Components.types';
-
-export type { HeroProps, EditorialProps, HowToProps };
