--- apps/nextjs-website/react-components/src/index.ts
@@ -1,5 +1,0 @@
-import Hero from './components/Hero/Hero';
-import Editorial from './components/Editorial/Editorial';
-import HowTo from './components/HowTo/HowTo';
-
-export { Hero, Editorial, HowTo };
