+++ apps/nextjs-website/react-components/types/index.ts
@@ -1,4 +1,5 @@
 import { HeroProps } from './Hero/Hero.types';
 import { EditorialProps } from './Editorial/Editorial.types';
+import { HowToProps } from './HowTo/HowTo.types';
 
-export type { HeroProps, EditorialProps };
+export type { HeroProps, EditorialProps, HowToProps };
