+++ apps/nextjs-website/src/components/Editorial.tsx
@@ -1,13 +1,12 @@
 'use client';
 import Image from 'next/image';
 import MarkdownRenderer from './MarkdownRenderer';
-import { Editorial as EditorialEC } from '@react-components';
-import { EditorialProps } from '@react-components-props';
+import { Editorial as EditorialRC } from '@react-components/components';
+import { EditorialProps } from '@react-components/types';
 import { EditorialSection } from '@/lib/fetch/types/PageSection';
 import Icon from '@/components/Icon';
 
 const makeEditorialProps = ({
-  theme,
   eyelet,
   body,
   image,
@@ -15,13 +14,13 @@ const makeEditorialProps = ({
   storeButtons,
   ...rest
 }: EditorialSection): EditorialProps => ({
-  theme,
   ...(eyelet && { eyelet }),
   body: MarkdownRenderer({ markdown: body, variant: 'body2' }),
   image: (
     <Image
-      src={`http://localhost:1337${image.url}`}
+      src={image.url}
       alt={image.alternativeText ?? ''}
+      // Needed by Image, will be overwritten
       width={0}
       height={0}
     />
@@ -43,7 +42,7 @@ const makeEditorialProps = ({
 });
 
 const Editorial = (props: EditorialSection) => (
-  <EditorialEC {...makeEditorialProps(props)} />
+  <EditorialRC {...makeEditorialProps(props)} />
 );
 
 export default Editorial;
